feat: banner slide admin controls, trust & safety support pages, PostgreSQL/Prisma migration
- Add tag, cta_text, product_label, and rating fields to banner_ads — admins can now manage all slide overlay content (badge text, CTA label, product label, star rating) from the dashboard
- Add four Trust & Safety support pages: /support/seller-verification, /support/escrow-policy, /support/buyer-guarantee, /support/dispute-resolution — fixing 404s from footer links
- Migrate database layer from MySQL to PostgreSQL via Prisma adapter (PrismaPg), with auto placeholder conversion (? → $N) and Decimal/BigInt/Date serialisation
- Add hero video and mega-menu admin features with corresponding API routes
- Update all API routes and SQL schemas for PostgreSQL compatibility

+++ src/app/api/admin/disputes/route.ts
@@ -34,10 +34,10 @@ export async function GET(request: NextRequest) {
     try {
       const summary = await query(`
         SELECT 
-          SUM(CASE WHEN status IN ('open', 'seller_responded') THEN 1 ELSE 0 END) as open_count,
-          SUM(CASE WHEN status = 'under_review' THEN 1 ELSE 0 END) as review_count,
-          SUM(CASE WHEN status IN ('resolved_buyer', 'resolved_seller', 'resolved_split', 'closed') THEN 1 ELSE 0 END) as resolved_count,
-          COUNT(*) as total
+          SUM(CASE WHEN status IN ('open', 'seller_responded')::float THEN 1 ELSE 0 END) as open_count,
+          SUM(CASE WHEN status = 'under_review' THEN 1 ELSE 0 END)::float as review_count,
+          SUM(CASE WHEN status IN ('resolved_buyer', 'resolved_seller', 'resolved_split', 'closed')::float THEN 1 ELSE 0 END) as resolved_count,
+          COUNT(*)::int as total
         FROM disputes
       `) as Record<string, unknown>[];
       if (summary.length > 0) {
